Refactor package name.
