+++ src/pycategory/__init__.py
