+++ src/pycategory/applicative.py
