+++ src/pycategory/collection.py
