+++ src/pycategory/constraints.py
