+++ src/pycategory/either.py
