+++ src/pycategory/extractor.py
