+++ src/pycategory/functor.py
