+++ src/pycategory/monad.py
