+++ src/pycategory/option.py
