+++ src/pycategory/pipeline.py
