+++ src/pycategory/processor.py
