+++ src/pycategory/try_.py
